requirements.txt
fastapi==0.78.0
uvicorn==0.18.2
requests==2.28.1
minio==7.1.9
Pillow==9.2.0
mysql-connector-python==8.0.29
typing-extensions==4.3.0